+++ CHANGELOG.md
@@ -6,7 +6,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 
 ### Changed
-- GetManagedDatabaseIndices method to return a slice of structs instead of pointers 
+- GetManagedDatabaseIndices method to return a slice of structs instead of pointers
+- ManagedDatabaseAccessControl struct fields to non-pointer values
 
 ## [6.2.0]