+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [6.2.1]
+
 ### Changed
 - GetManagedDatabaseIndices method to return a slice of structs instead of pointers 
 
@@ -288,7 +290,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.2.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.2.1...HEAD
+[6.2.1]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.2.0...v6.2.1
 [6.2.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.1.1...v6.2.0
 [6.1.1]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.1.0...v6.1.1
 [6.1.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v6.0.0...v6.1.0